fix(core): Forensic audit fixes for C# stability, ML integration, and data integrity
- Fixed async deadlocks in Program.cs and MiniAppController

- Eliminated Train-Serve Skew by aligning Python and C# ISO-8601 timestamps

- Fixed _get_sec parser crashing on subminute string intervals

- Fixed Doji noise filtering and ML feedback loop in TradeOutcomeTracker

- Restored local PostgreSQL database connectivity in start_ml.bat and run_with_env

- Patched security bypasses in Telegram validation and AuthService

- Fixed backtester.py syntax error and verify_data.py Docker path hardcoding

+++ MiniApp/Data/Repositories/RegistrationRepository.cs
@@ -9,11 +9,11 @@ namespace ValutaBot.App.MiniApp.Data.Repositories
 {
     public static class RegistrationRepository
     {
-        public static async Task<int> GetRegistrationsCountAsync()
+        public static async Task<long> GetRegistrationsCountAsync()
         {
             if (string.IsNullOrEmpty(DbConnectionFactory.GetConnectionString())) return 0;
             using var conn = DbConnectionFactory.GetConnection();
-            return (int)await conn.ExecuteScalarAsync<long>("SELECT COUNT(*) FROM registrations");
+            return await conn.ExecuteScalarAsync<long>("SELECT COUNT(*) FROM registrations");
         }
 
         public static async Task<List<TelegramBotService.PocketRegistration>> GetLatestRegistrationsAsync(int limit = 15)
